move SessionSettings into its own package
