+++ src/org/tn5250j/settings/OnConnectAttributesPanel.java
@@ -31,7 +31,7 @@ import javax.swing.*;
 import org.tn5250j.tools.*;
 import org.tn5250j.SessionConfig;
 
-public class OnConnectAttributesPanel extends AttributesPanel {
+ class OnConnectAttributesPanel extends AttributesPanel {
 
    private static final long serialVersionUID = 1L;
 JTextField connectMacro;
